services/umap-services/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.104.1
h11==0.16.0
idna==3.11
joblib==1.5.3
llvmlite==0.46.0
numba==0.64.0
numpy==1.24.3
pydantic==2.5.0
pydantic_core==2.41.5
pynndescent==0.6.0
scikit-learn==1.7.2
scipy==1.15.3
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
umap-learn==0.5.5
uvicorn[standard]==0.24.0
